+++ src/models/Entry.php
@@ -6,6 +6,7 @@ namespace davidhirtz\yii2\cms\tenant\models;
 
 use davidhirtz\yii2\cms\models\queries\EntryQuery;
 use davidhirtz\yii2\cms\tenant\behaviors\EntryTenantBehavior;
+use davidhirtz\yii2\tenant\models\Tenant;
 use Yii;
 
 /**
@@ -30,6 +31,12 @@ class Entry extends \davidhirtz\yii2\cms\models\Entry
                 ...$this->getEntryTenantBehavior()->getTenantRouteParams()
             ]
             : false;
+    }
+
+    public function populateTenantRelation(Tenant $tenant): void
+    {
+        $this->populateRelation('tenant', $tenant);
+        $this->tenant_id = $tenant->id;
     }
 
     public function findSiblings(): EntryQuery